レビュー機能完成

+++ database/migrations/2023_07_30_005832_create_reviews_table.php
@@ -16,7 +16,7 @@ class CreateReviewsTable extends Migration
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
             $table->foreignId('reserve_id')->unique()->constrained('reserves');
-            $table->integer('star')->unsigned()->nullable(false);
+            $table->double('star',2,1)->unsigned()->nullable(false);
             $table->string('comment', 500)->nullable(true);
             $table->timestamp('created_at')->useCurrent()->nullable();
             $table->timestamp('updated_at')->useCurrent()->nullable();
